tema3/requirements.txt
paho-mqtt==1.5.0
influxdb-client==1.18.0